move the new method inside com/microfocus/application/automation/tools/JenkinsUtils.java
